Triangle. Closed #25.

+++ triangle/Triangle.java
@@ -5,9 +5,9 @@ public class Triangle{
         Scanner scanner = new Scanner(System.in);
         double base, height, ans;
         System.out.println("What is the base of a triangle? ");
-        base = scanner.nextInt();
+        base = scanner.nextDouble();
         System.out.println("What is the height of a triangle? ");
-        height = scanner.nextInt();
+        height = scanner.nextDouble();
         ans = base * height/2;
         System.out.printf("The Area Of A Triangle is %2f. ", ans);
